+++ webpack_boilerplate/frontend_template/{{cookiecutter.project_slug}}/src/application/app.js
@@ -1,13 +1,14 @@
 // This is the scss entry file
 import "../styles/index.scss";
 
-// We can import Bootstrap JS instead of the CDN link, if you do not use
-// Bootstrap, please feel free to remove it.
-import "bootstrap/dist/js/bootstrap.bundle";
-
 // We can import other JS file as we like
-import "../components/sidebar";
+import Jumbotron from "../components/jumbotron";
 
 window.document.addEventListener("DOMContentLoaded", function () {
-  window.console.log("dom ready 1");
+  window.console.log("dom ready");
+
+  // Find elements and initialize
+  for (const elem of document.querySelectorAll(Jumbotron.selector())) {
+    new Jumbotron(elem);
+  }
 });