+++ webpack_boilerplate/frontend_template/{{cookiecutter.project_slug}}/src/components/jumbotron.js
@@ -1,0 +1,13 @@
+class Jumbotron {
+  static selector() {
+    return "[data-jumbotron]";
+  }
+
+  constructor(node) {
+    this.node = node;
+    console.log(`Jumbotron initialized for node: ${node}`);
+    // do something here
+  }
+}
+
+export default Jumbotron;