+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## v1.0.3
+
+1. Remove `Bootstrap` from the default setup.
+2. Update doc about `Bootstrap`, `SWC`, `React`, `Vue`
+6. Update frontend dependency to the latest version
+
 ## v1.0.2
 
 1. Support Node LTS Iron